requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.6
django-bootstrap4==2.3.1
django-phone-field==1.8.1
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
soupsieve==2.1
sqlparse==0.4.1
